refactor: move metrics tracker to shared utils
Relocate createMetricsTracker from src/plugin/src/reporting/metrics.js to src/shared/utils/metrics.js so the instrumentation helper sits in the shared layer for reuse across subsystems.
