+++ src/plugin/src/reporting/index.js
@@ -1,1 +1,1 @@
-export * from "./metrics.js";
+export * from "../../../shared/metrics-utils.js";
