+++ src/plugin/tests/project-index-metrics.test.js
@@ -5,7 +5,7 @@ import path from "node:path";
 import test from "node:test";
 
 import { buildProjectIndex } from "../src/project-index/index.js";
-import { createMetricsTracker } from "../src/reporting/metrics.js";
+import { createMetricsTracker } from "../src/reporting.js";
 
 async function writeProjectFile(rootDir, relativePath, contents) {
     const absolutePath = path.join(rootDir, relativePath);
