+++ src/shared/utils/metrics.js
@@ -1,8 +1,5 @@
-import {
-    getNonEmptyString,
-    normalizeStringList
-} from "../../../shared/string-utils.js";
-import { toArrayFromIterable } from "../../../shared/array-utils.js";
+import { getNonEmptyString, normalizeStringList } from "./string.js";
+import { toArrayFromIterable } from "./array.js";
 
 const hasHrtime = typeof process?.hrtime?.bigint === "function";
